move long running tests that read/write from disk
