+++ tests/unit/lib/test_datachain.py
@@ -1423,47 +1423,6 @@ def test_explode_raises_on_wrong_column_type(test_session):
         dc.explode("f1.count")
 
 
-# These deprecation warnings occur in the datamodel-code-generator package.
-@pytest.mark.filterwarnings("ignore::pydantic.warnings.PydanticDeprecatedSince20")
-def test_to_from_json(tmp_dir, test_session):
-    df = pd.DataFrame(DF_DATA)
-    dc_to = DataChain.from_pandas(df, session=test_session)
-    path = tmp_dir / "test.json"
-    dc_to.order_by("first_name", "age").to_json(path)
-
-    with open(path) as f:
-        values = json.load(f)
-    assert values == [
-        {"first_name": n, "age": a, "city": c}
-        for n, a, c in zip(DF_DATA["first_name"], DF_DATA["age"], DF_DATA["city"])
-    ]
-
-    dc_from = DataChain.from_json(path.as_uri(), session=test_session)
-    df1 = dc_from.select("json.first_name", "json.age", "json.city").to_pandas()
-    df1 = df1["json"]
-    assert df_equal(df1, df)
-
-
-# These deprecation warnings occur in the datamodel-code-generator package.
-@pytest.mark.filterwarnings("ignore::pydantic.warnings.PydanticDeprecatedSince20")
-def test_from_json_jmespath(tmp_dir, test_session):
-    df = pd.DataFrame(DF_DATA)
-    values = [
-        {"first_name": n, "age": a, "city": c}
-        for n, a, c in zip(DF_DATA["first_name"], DF_DATA["age"], DF_DATA["city"])
-    ]
-    path = tmp_dir / "test.json"
-    with open(path, "w") as f:
-        json.dump({"author": "Test User", "version": 5, "values": values}, f)
-
-    dc_from = DataChain.from_json(
-        path.as_uri(), jmespath="values", session=test_session
-    )
-    df1 = dc_from.select("values.first_name", "values.age", "values.city").to_pandas()
-    df1 = df1["values"]
-    assert df_equal(df1, df)
-
-
 def test_to_json_features(tmp_dir, test_session):
     dc_to = DataChain.from_values(
         f1=features, num=range(len(features)), session=test_session
